move value to manager
